yakuzai_trainer/requirements.txt
annoy==1.17.1
asgiref==3.5.2
azure-cognitiveservices-vision-computervision==0.9.0
azure-cognitiveservices-vision-customvision==3.1.0
azure-common==1.1.28
azure-core==1.25.1
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1
django-forms-bootstrap==3.1.0
idna==3.4
ImageHash==4.3.0
isodate==0.6.1
msrest==0.7.1
numpy==1.23.3
oauthlib==3.2.1
opencv-python==4.6.0.66
Pillow==9.2.0
psycopg2==2.9.3
python-dotenv==0.21.0
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.1
scipy==1.9.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
